refactor(telemetry): extract shared robot recording into hud.telemetry.robot
Recording is shared infrastructure (agent harness, hud.wrap, gym bridges),
not agent code. TraceRecorder (one trace, span emission only - lifecycle
stays with the caller) and JobRecorder (a vectorized env as one Job of
per-episode traces) replace Recorder/VecRecorder/EpisodeRecorder and their
dual-mode flags; video streaming moves alongside. The LeRobot dataset half
of record.py becomes agents/robot/dataset.py (DatasetWriter). InferenceStep
gains contract-derived per-dim action names so the viewer can label plots.
