+++ hud/__init__.py
@@ -8,7 +8,6 @@ from __future__ import annotations
 # Apply patches to third-party libraries early, before other imports
 from . import patches as _patches  # noqa: F401
 from ._legacy import install as _install_v5_compat
-from .agents.robot.record import Recorder, VecRecorder
 from .clients import connect
 from .environment import Environment
 from .eval import (
@@ -46,7 +45,6 @@ __all__ = [
     "HostedRuntime",
     "Job",
     "LocalRuntime",
-    "Recorder",
     "Run",
     "Runtime",
     "RuntimeConfig",
@@ -59,7 +57,6 @@ __all__ = [
     "Taskset",
     "Trace",
     "TrainingClient",
-    "VecRecorder",
     "__version__",
     "connect",
     "instrument",
